高德地图MCP_cursor/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
jieba==0.42.1
